Finalização do Back-end

+++ Backend/controllers/DatabaseController.js
@@ -1,21 +1,7 @@
 const Database = require("../models/Database");
+const ImageType = require("../models/ImageType");
 
 class DatabaseController {
-    async newDatabase (req, res) {
-        const { name, examType, description, imageQuality, imageType, sourceLink } = req.body;
-
-        if(await Database.databaseExists(name)){
-            res.status(403).json({ err: "Já existe banco de dados cadastrado com esse nome!" });
-            return;
-        }
-
-        const result = await Database.create(name, examType, description, imageQuality, imageType, sourceLink);
-
-        if (result.sucess) {
-            res.status(201).json({ msg: "Banco de dados criado com sucesso!." });
-            return
-        }
-    }
 
     async getDatabases (req, res) {
         const databases = await Database.find();
@@ -26,42 +12,91 @@ class DatabaseController {
 
     async getDatabase (req, res) {
         const id = req.params.id;
+        const databases = await Database.find({ _id: id });
 
-        const database = await Database.find({ _id: id });
+        if (databases[0]){
+            res.status(200).json(databases[0]);
+            return
+        } else {
+            res.status(404).json({ err: "Banco de dados não encontrado" });
+            return;
+        }
+    }
 
-        res.status(200).json(database[0]);
-        return
+    async newDatabase (req, res) {
+        const { name, examType, description, imageQuality, imageType, sourceLink } = req.body;
+
+        if(await Database.databaseExists(name)){
+            res.status(403).json({ err: "Já existe banco de dados cadastrado com esse nome!" });
+            return;
+        }
+
+        if (!await ImageType.imageTypeExists(imageType)){
+            res.status(400).json({ err: "Tipo de imagem não existe!" });
+            return;
+        }
+
+        const result = await Database.create(name, examType, description, imageQuality, imageType, sourceLink);
+
+        if (result.sucess) {
+            res.status(201).json({ msg: "Banco de dados criado com sucesso!." });
+            return
+            
+        } else {
+            res.status(400).json({ err: "Erro ao cadastrar." });
+            return;
+        }
     }
 
     async updateDatabase (req, res) {
         const id = req.params.id;
         const { name, examType, description, imageQuality, imageType, sourceLink } = req.body;
+        const database = await Database.find({ _id: id });
 
-        const result = await Database.update(id, name, examType, description, imageQuality,
-                                                imageType, sourceLink);
-        
-        if (result.sucess){
+        if (!database[0]) {
+            res.status(404).json({ err: "Banco de dados não encontrado!" });
+            return;
+        }
+
+        if (!await ImageType.imageTypeExists(imageType)){
+            res.status(400).json({ err: "Tipo de imagem não existe!" });
+            return;
+        }
+
+        const databaseExists = await Database.databaseExists(name);
+        const namesEquals = name.toLowerCase() && database[0].name;
+
+        if (!databaseExists || databaseExists && namesEquals){
+            await Database.update(id, name, examType, description, imageQuality, imageType, sourceLink);
+
             res.status(200).json({ msg: "Banco de dados atualizado com sucesso" });
             return;
+
         } else {
-            res.status(404).json({ err: "Banco de dados não encontrado" });
-            return
-        }
+            res.status(403).json({ err: "Já existe banco de dados cadastrado com esse nome!" });
+            return;
+        }  
     }
 
     async deleteDatabase (req, res) {
         const { id } = req.params;
+        const database = await Database.find({ _id: id });
 
-        try {
-            const result = await Database.delete(id);
-
-            if (result.sucess) {
+        if (database[0]) {
+            try {
+                await Database.delete(id);
+    
                 res.status(200).json({ msg: "Banco de dados deletado com sucesso!" });
-            } else {
-                res.status(404).json({ msg: "Banco de dados não encontrado!" });
+                return;
+                
+            } catch (error) {
+                res.status(400).json({ msg: "Erro ao deletar!." });
+                return;
             }
-        } catch (error) {
-            res.status(400).json({ msg: "Erro ao deletar!." });
+
+        } else {
+            res.status(404).json({ err: "Banco de dados não encontrado!" });
+            return;
         }
     }
 }
